Rewrite rendering flow to use glsl
Some nodes were removed completely in the process, either because of
large dependencies or just because they turned out to not be useful.
Specifically, SobelNode was removed only because of the external
dependencies, so it should be reimplemented in glsl by someone.

Rendering is now deferred to a separate render loop to improve
performance and UI snappiness.